requirements.txt
# Core Dependencies
python-telegram-bot==20.7
pyyaml>=6.0.1
aiohttp>=3.9.1
python-dotenv>=1.0.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Production (optional)
# redis>=5.0.0  # For distributed memory
# sqlalchemy>=2.0.0  # For database storage
# schedule>=1.2.0  # For advanced scheduling
